bid-analyzer/requirements.txt
psycopg2-binary
pandas
sqlalchemy
scikit-learn
joblib
python-dotenv
